added folders for teller and base user
